Initial commit.
Basically this is ed ripped off from p9p;
the idea is to break the build to get it to (cross)compile with mingw
and try github while I struggle with win32.

While I don't think anyone should be interested in this project,
if you are reading this let me warn you: THE NEXT 1e9 COMMITS WILL
SIMPLY *NOT* COMPILE -- getting it to compile using the winapi is one of
the goals!