notify.c: sigaction->signal
Based on MSDN it seems that signal(2) on windows is reasonable,
ie handlers are forever.

I have no idea how(if?) to address the issues with sigmasks;
I've commented it out atm, but I'm also adding a hook to remind me of
any FIXME/HACK/TODO/XXX in the code.