Cross compiling requires cross-ar(1). D'oh!
The way it is now it's a horrible, horrible kludge:
get compile.XYZ that works, copy it to ./compile, and use the
corresponding ar.XYZ for libraries.

So brittle and ugly!

A cleaner way to do it
======================

Have many cross.XYZ files, each containing a line like

CROSS=mingwxyz-

try $CROSS, and then each of this files in turn, until $CROSS$CC works;
then write both compile and ar.