Fixed!
This really made me want to scream; see if you can understand the
problem from the diff + scratch/Xjmp.c (nvm the preprocessor).

It still needs a bit of real-windows testing anyway.

PS
My keyboard crashed while I was editing this message.
Yep, my keyboard -- everything else was working.
Using vim+spectrwm :D

Solution
========

You cannot return after a setjmp and expect the jmp_buf to work.
What makes me sad is how long it took me to realize it -- d'oh!
The preprocessor lines in the diff are misleading -- they only exist to
cope with the p9*jmp macros, exploiting the fact that on mingw setjmp() is
an alias for _setjmp().

Btw, I'll have to make this sort of a recursive commit, because this
things aren't commented enough in u.h/libc.h...